refactor: 세르파 정보 컴포넌트 분리
